added styling to the list-element
